Update fastFit.m
made fast fit have useDefaults option which decides if the user wants to use the default values. If there are not enough arguments, ask the user if they want to use the default values. They also have the option to just call fastFit with only x, y, figNum, useDefaults. If they want to use the defaults, the default values will be inserted. If they don't they will be asked to input all of the stuff they didn't already give.